requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
lxml>=4.9.3
python-dotenv>=1.0.0
streamlit>=1.24.0
plotly>=5.13.0
jupyter>=1.0.0
matplotlib>=3.7.1
seaborn>=0.12.2 